refactor: move diff to utils
